casting lists

+++ lib/src/resources/news_api_provider.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:async';
 
 import 'package:http/http.dart' show Client;
 import '../models/item_model.dart';
@@ -11,7 +12,7 @@ class NewApiProvider {
   Future<List<int>> fetchTopIds() async {
     final response = await client.get(Uri.parse('$_root/topstories.json'));
     final ids = json.decode(response.body);
-    return ids;
+    return ids.cast<int>();
   }
 
   Future<ItemModel> fetchItem(int id) async {
